Updated to IBPP version 2.4.4.0
svn path=/trunk/flamerobin/; revision=569

+++ src/ibpp/_epb.cpp
@@ -121,7 +121,7 @@ void EPB::FireActions(IBPP::IDatabase* db)
 		{ // Fire the action
 			try
 			{
-				(*ref_it)->ibppEventHandler(db, event_buffer_it.get_name(), vnew - vold);
+				(*ref_it)->ibppEventHandler(db, event_buffer_it.get_name(), (int)(vnew - vold));
 			} catch (...) { }
 		}
 	}
